Create Compare The Triplets Hackerrank Solution C++
Compare The Triplets Hackerrank Solution C++. Alice and Bob each created one problem for HackerRank. A reviewer rates the two challenges, awarding points on a scale from 1 to 100 for three categories: problem clarity, originality, and difficulty. Triplets Hackerrank Solution. The rating for Alice's challenge is the triplet a = (a[0], a[1], a[2]), and the rating for Bob's challenge is the triplet b = (b[0], b[1], b[2]).


Compare The Triplets problem similar to comparing marks of three subjects of two students in a way like if subject1, student1 got high marks than student2 then student1 got 1 point and the same repeat for subject2, and subject3 at the end print both student points. The maximum number of points can a student earn is 3 or the minimum can be 0 or there may be an equal point if subject1 both students got the same point.
The task is to find their comparison points by comparing a[0] with b[0], a[1] with b[1], and a[2] with b[2].

If a[i] > b[i], then Alice is awarded 1 point.
If a[i] < b[i], then Bob is awarded 1 point.
If a[i] = b[i], then neither person receives a point.

Comparison points are the total points a person earned.


Given a and b, determine their respective comparison points.

Example

a = [1, 2, 3]
b = [3, 2, 1]

For elements *0*, Bob is awarded a point because of a[0].
For the equal elements a[1] and b[1], no points are earned.
Finally, for elements 2, a[2] > b[2] so Alice receives a point.

The returned array is [1, 1] with Alice's score first and Bob's second.


Function Description

Complete the function compare Triplets in the editor below.

compare 


Triplets has the following parameter(s):

int a[3]: Alice's challenge rating
int b[3]: Bob's challenge rating

Return

int[2]: Alice's score is in the first position, and Bob's score is in the second.

Input Format


The first line contains 3 space-separated integers, a[0], a[1], and a[2], the respective values in triplet a.
The second line contains 3 space-separated integers, b[0], b[1], and b[2], the respective values in triplet b.

Constraints

1 ≤ a[i] ≤ 100
1 ≤ b[i] ≤ 100

Sample Input 0

5 6 7
3 6 10


Sample Output 0

1 1